Rationalising the number of deterministic clocks in our test code.
